pyproject.toml
[project]
name = "omerocrate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Michael Milton", email = "ttmigueltt@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "mypy",
    "pyyaml>=6.0.2",
    "rdflib>=7.1.1",
    "typer>=0.15.2",
]

[project.scripts]
omerocrate = "omerocrate.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
stubPath = "./typings"

[tool.uv.sources]
mypy = { git = "https://github.com/python/mypy" }

[dependency-groups]
dev = [
    "dotenv>=0.9.9",
    "gitpython>=3.1.44",
    "mkdocs-same-dir>=0.1.3",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.7",
    "mkdocstrings[python]>=0.28.3",
    "pytest>=8.3.5",
    "mkdocs-typer>=0.0.3",
]
